[lldb][NFC] Move VectorIterator to CPlusPlusLanguage
The VectorIterator header was still in DataFormatter so I moved it.
While I was there, I also moved the implementation to its own file to
mimic other SyntheticChildrenFrontEnds (e.g. Coroutines).
