+++ lldb/source/Plugins/Language/CPlusPlus/MsvcStlVector.cpp
@@ -8,9 +8,10 @@
 
 #include "MsvcStl.h"
 
+#include "Plugins/Language/CPlusPlus/VectorIterator.h"
+
 #include "lldb/DataFormatters/FormattersHelpers.h"
 #include "lldb/DataFormatters/TypeSynthetic.h"
-#include "lldb/DataFormatters/VectorIterator.h"
 #include "llvm/Support/ErrorExtras.h"
 
 using namespace lldb;
